Create Fee-Structure
contract constructor sets a deposit fee of 0.01 ETH for the life of each Ethscription. Users need to send at least this amount when depositing an Ethscription.

The makePayment function allows users to send payments to the wrapped address associated with their Ethereum wallet. The function checks if a wrapped address exists for the sender and emits a PaymentReceived event with details about the sender, receiver, and payment amount.